requirements.txt
# Python 3.12.1
# Node 20.11.1
# Vue 3.4.21
asgiref==3.7.2
distlib==0.3.8
Django==5.0.3
djangorestframework==3.14.0
filelock==3.13.1
mysqlclient==2.2.4
platformdirs==4.2.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
